pyproject.toml
[project]
name = "tab-organizer"
version = "0.1.0"
description = "Scripts for Tab Organizer Chrome extension"
requires-python = ">=3.13"
dependencies = [
  "pyobjc-framework-Quartz>=12.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
